refactor to single instance of dialog
